Move to nextjs 12.2
